% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ambi.R
\name{compute_ambi}
\alias{compute_ambi}
\title{Compute AMBI index for a dataset}
\usage{
compute_ambi(data, taxon, count, group)
}
\arguments{
\item{data}{Data frame with taxon, count, and optionally group}

\item{taxon}{Column name for taxa}

\item{count}{Column name for abundance}

\item{group}{Column name for ecological group (EG I–V)}
}
\value{
AMBI index (numeric)
}
\description{
Compute AMBI index for a dataset
}
