% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_img_crop.R
\name{fct_img_crop}
\alias{fct_img_crop}
\title{Crop map images}
\usage{
fct_img_crop(path)
}
\arguments{
\item{path}{file path}
}
\value{
a cropped map to be used in cowplot::plot_grid()
}
\description{
Delete empty space around map ggplots
}
