% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_openstreet_base_map.R
\name{fct_openstreet_base_map}
\alias{fct_openstreet_base_map}
\title{Create OpenStreetMap base map for a given region}
\usage{
fct_openstreet_base_map(data)
}
\arguments{
\item{data}{}
}
\value{
1
}
\description{
Create OpenStreetMap base map for a given region
}
