% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_size_class.R
\name{fct_size_class}
\alias{fct_size_class}
\title{fct_size_class Title}
\usage{
fct_size_class(data, species)
}
\arguments{
\item{data}{}

\item{species}{}
}
\value{
1
}
\description{
fct_size_class Title
}
