% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_bbox_aspect.R
\name{fix_bbox_aspect}
\alias{fix_bbox_aspect}
\title{Title}
\usage{
fix_bbox_aspect(xlim, ylim, target_ratio = 1)
}
\arguments{
\item{xlim}{}

\item{ylim}{}

\item{target_ratio}{}
}
\description{
Description
}
