% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_seg_model.R
\name{fun_seg_model}
\alias{fun_seg_model}
\title{Title}
\usage{
fun_seg_model(
  data,
  estuary,
  chemical,
  folder = "metals",
  n_break = 1,
  plot_save = TRUE
)
}
\description{
Description
}
\examples{
# fun_seg_model()
}
