% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_estuary_from_gps_position.R
\name{get_estuary_from_gps_position}
\alias{get_estuary_from_gps_position}
\title{Get estuary variable from GPS positions}
\usage{
get_estuary_from_gps_position(data, latitude, longitude)
}
\arguments{
\item{data}{tibble containing columns with GPS position through latitude and longitude columns}

\item{latitude}{name of latitude column}

\item{longitude}{name of longitude column}
}
\value{
the tibble + variable estuary
}
\description{
Get estuary variable from GPS positions
}
\examples{
toy_data_gps <- data.frame(
  station = c(
    "GIR_01", "GIR_02", "GIR_03",
    "LOI_01", "LOI_02", "LOI_03",
    "SEI_01", "SEI_02", "SEI_03"
  ),
  lat = c(
    45.58, 45.52, 45.45,
    47.28, 47.30, 47.25,
    49.47, 49.45, 49.42
  ),
  lon = c(
    -0.95, -1.05, -1.15,
    -2.05, -2.20, -2.35,
    0.18, 0.30, 0.42
  )
)

get_estuary_from_gps_position(data = toy_data_gps, latitude = lat, longitude = lon)
}
