% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_general_evaluation.R
\name{get_general_evaluation}
\alias{get_general_evaluation}
\title{Title}
\usage{
get_general_evaluation(family, suffix = "", res_family_trends, res_kruskal)
}
\arguments{
\item{family}{character Family folder name}

\item{suffix}{character If necessary, a suffix for the output .xlsx file name}

\item{res_family_trends}{tibble}

\item{res_kruskal}{tibble}
}
\value{
save a .xlsx in the folder "../inst/results/data_contam/"
}
\description{
Description
}
\examples{
# get_general_evaluation()
}
