% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_general_levels.R
\name{get_general_levels}
\alias{get_general_levels}
\title{Get general levels in the original unit of measurment}
\usage{
get_general_levels(data, family, suffix)
}
\arguments{
\item{data}{tibble with ESTUARY, PARAMETRE_LIBELLE, periode, RESULTAT}

\item{family}{character Family folder name}

\item{suffix}{character If necessary, a suffix for the output .xlsx file name}
}
\value{
save a .xlsx in the folder "../inst/results/data_contam/"
}
\description{
Get general levels in the original unit of measurment
}
\examples{
# get_general_levels()
}
