% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info_from_dates.R
\name{get_info_from_dates}
\alias{get_info_from_dates}
\title{Get information from dates}
\usage{
get_info_from_dates(data, date_variable)
}
\arguments{
\item{data}{tibble containing a column of class date}

\item{date_variable}{name of the date column}
}
\value{
the tibble + variables year, month, year_month, season, and year_season
}
\description{
Get year, month, year_month, season, and year_season variables from a date variable
}
\examples{
# data("economics")
# get_info_from_dates(data = economics, date_variable = date)
}
