% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kruskal_periods_results.R
\name{get_kruskal_periods_results}
\alias{get_kruskal_periods_results}
\title{Get results of Kruskal-Wallis test between two periods of time}
\usage{
get_kruskal_periods_results(
  data_with_periode,
  chemicals,
  estuaries = c("Gironde", "Loire", "Seine")
)
}
\arguments{
\item{data_with_periode}{Dataset with columns ESTUARY, PARAMETRE_LIBELLE, RESULTAT, periode}

\item{chemicals}{vector of existing values in PARAMETRE_LIBELLE}

\item{estuaries}{vector of existing values in ESTUARY}
}
\value{
tibble containing, ESTUARY, PARAMETRE_LIBELLE, pvalue, median_1 (first period), median_2 (last period), max_value, trend (symbol to be plotted), y_pos (y position to be plotted)
}
\description{
Get all information of Kruskal-Wallis test by PARAMETRE_LIBELLE and ESTUARY to be plotted
}
\examples{
# get_kruskal_periods_results()
}
