% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_last_trend_info.R
\name{get_last_trend_info}
\alias{get_last_trend_info}
\title{Get results from Spearman's test for several estuaries and contaminant}
\usage{
get_last_trend_info(data, norm)
}
\arguments{
\item{data}{tibble with ESTUARY, PARAMETRE_LIBELLE, YEAR, RESULTAT}

\item{norm}{tibble PARAMETRE_LIBELLE, NORM}
}
\value{
tibble ESTUARY, PARAMETRE_LIBELLE, rho, p.value, symbol, last_year, last_resultat, norm
}
\description{
Last 15 years trend and significance
}
\examples{
# get_last_trend_info()
}
