% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_map_parametre_data_benthos.R
\name{gg_map_parametre_data_benthos}
\alias{gg_map_parametre_data_benthos}
\title{Title}
\usage{
gg_map_parametre_data_benthos(data, estuaire, color)
}
\arguments{
\item{data}{}

\item{estuaire}{}

\item{color}{}
}
\value{
ggplot
}
\description{
Description
}
\examples{
# gg_map_parametre_data_benthos()
}
