% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_colors.R
\name{ggplot2_colors}
\alias{ggplot2_colors}
\title{ggplot2 colors}
\usage{
ggplot2_colors(n)
}
\arguments{
\item{n}{number of colors needed}
}
\value{
vector of hexadecimal colors
}
\description{
Reproduce a vector of colors like those from {ggplot2}
}
\examples{
ggplot2_colors(1)
}
