% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_maps_parameter_years.R
\name{plot_maps_parameter_years}
\alias{plot_maps_parameter_years}
\title{Map yearly mean GPS position by estuary and haline zone}
\usage{
plot_maps_parameter_years(data, parameter, filename)
}
\arguments{
\item{data}{tibble, data containing PARAMETRE_LIBELLE, estuary, year, haline_zone, latitude, longitude}

\item{parameter}{character, name of the parameter GPS measurment positions to be mapped}

\item{filename}{character, complete path where to save the map}
}
\value{
saved ggplot
}
\description{
Map yearly mean GPS position by estuary and haline zone
}
