% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trends_by_parameter.R
\name{plot_trends_by_parameter}
\alias{plot_trends_by_parameter}
\title{Automate graph function for time-series across all parameters}
\usage{
plot_trends_by_parameter(
  data,
  gam_results,
  response_col = "RESULTAT",
  x_col = "year_month",
  parameter_col = "PARAMETRE_LIBELLE",
  grouping = NULL,
  colour_var = NULL,
  facet_rows = NULL,
  facet_cols = NULL,
  auto_layout = TRUE,
  ...
)
}
\value{
list of ggplots
}
\description{
Description
}
