% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gam_by_parameter.R
\name{run_gam_by_parameter}
\alias{run_gam_by_parameter}
\title{Title}
\usage{
run_gam_by_parameter(
  data,
  response_col = "RESULTAT",
  parameter_col = "PARAMETRE_LIBELLE",
  year_col = "year",
  group_vars = c("estuary", "haline_zone", "season"),
  all_group_combinations = FALSE,
  k_max = 6,
  min_n = 6,
  min_years = 6
)
}
\arguments{
\item{data}{}

\item{response_col}{character, name of the column containing measurements values}

\item{year_col}{character, name of the column containing years}

\item{group_vars}{vector of characters, groupin variables}

\item{all_group_combinations}{logical, TRUE test all combinations of the different variables from group_vars, FALSE test only with the combination of the group_vars given}

\item{k_max}{}

\item{min_n}{}

\item{min_years}{}
}
\description{
Description
}
