% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spearman_test.R
\name{spearman_test}
\alias{spearman_test}
\title{Get results from Spearman's test}
\usage{
spearman_test(x, y)
}
\arguments{
\item{x}{first variable}

\item{y}{second variable}
}
\value{
tibble rho and p.value from Spearman's test
}
\description{
Save rho and p.value from Spearman's test in a tibble
}
\examples{
# spearman_test()
}
