% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_build_and_save_basemap.R
\name{fct_build_and_save_basemap}
\alias{fct_build_and_save_basemap}
\title{Create and save OpenStreetMap base map for an estuary}
\usage{
fct_build_and_save_basemap(
  data,
  estuary_name,
  villes_selection = NULL,
  path = "inst/extdata/basemaps"
)
}
\arguments{
\item{data}{Data frame containing latitude and longitude}

\item{estuary_name}{Character name of estuary (used for folder name)}

\item{villes_selection}{Character vector of city names to keep (optional)}

\item{path}{Base path for saving (default: inst/extdata/basemaps)}
}
\value{
A list with map components (also saved on disk)
}
\description{
Create and save OpenStreetMap base map for an estuary
}
\examples{
# fct_build_and_save_basemap(
#   data = esteem.overview::data_POMET_ALL_densities |> filter(estuary == "Gironde"),
#   estuary_name = "Gironde",
#   villes_selection = c("Royan", "Pauillac", "Saint-Estèphe", "Blaye", "Saint-Christoly-Médoc")
# )
}
