% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_formatting_raw_data_pomet.R
\name{fct_formatting_raw_data_POMET}
\alias{fct_formatting_raw_data_POMET}
\title{Formatting POMET raw datasets}
\usage{
fct_formatting_raw_data_POMET(data_POMET)
}
\arguments{
\item{data_POMET}{}
}
\value{
tibble
}
\description{
Formatting POMET raw datasets
}
