% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_map_parametre_data_benthos.R
\name{gg_map_parametre_data_benthos}
\alias{gg_map_parametre_data_benthos}
\title{Get map of REBENT sampling points for each year of monitoring.}
\usage{
gg_map_parametre_data_benthos(data, estuaire)
}
\arguments{
\item{data}{tibble data benthos with ESTUARY, YEAR, LIEU_MNEMONIQUE, latitude, longitude}

\item{estuary}{character "Gironde", "Loire", or "Seine"}

\item{color}{character color}
}
\value{
ggplot 1 map per year
}
\description{
Get map of REBENT sampling points for each year of monitoring.
}
\examples{
# gg_map_parametre_data_benthos()
}
