% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_estuary_basemap.R
\name{load_estuary_basemap}
\alias{load_estuary_basemap}
\title{Load estuary basemap (OpenStreetMap)}
\usage{
load_estuary_basemap(estuary_name, package = "esteem.overview")
}
\arguments{
\item{estuary_name}{Character name of estuary}

\item{package}{Name of package (default = "esteem.overview")}
}
\value{
List containing osm_raster, villes, xlim, ylim
}
\description{
Load estuary basemap (OpenStreetMap)
}
\examples{
basemap <- load_estuary_basemap(estuary_name = "Loire" )
# 
# ggplot() +
#   tidyterra::geom_spatraster_rgb(data = basemap$osm_raster) +
#   geom_sf(data = basemap$villes) +
#   coord_sf(xlim = basemap$xlim,
#            ylim = basemap$ylim,
#            expand = FALSE)
}
