% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_estuary_map.R
\name{plot_estuary_map}
\alias{plot_estuary_map}
\title{Plot estuary map with points, optional variable mapping and optional horizontal/vertical lines}
\usage{
plot_estuary_map(
  data,
  estuary_name,
  colour_var,
  fill_var,
  size_var,
  shape_var,
  package = "esteem.overview"
)
}
\arguments{
\item{data}{Data frame containing latitude and longitude}

\item{estuary_name}{Character name of the estuary}

\item{colour_var}{Optional variable in \code{data} for point colour}

\item{fill_var}{Optional variable in \code{data} for point fill (shape 21)}

\item{size_var}{Optional variable in \code{data} for point size}

\item{shape_var}{Optional variable in \code{data} for point shape}

\item{package}{Package name where basemap is stored (default: "esteem.overview")}
}
\value{
ggplot object
}
\description{
Plot estuary map with points, optional variable mapping and optional horizontal/vertical lines
}
\examples{
# plot_estuary_map(
#   data = data_POMET |> filter(estuary == "Gironde"),
#   estuary_name = "Gironde",
#   colour_var = salinite,
# )
}
